Fixed the calculation of average response time per interaction (again).
git-svn-id: https://osdldbt.svn.sourceforge.net/svnroot/osdldbt/trunk/dbt1@40 eccdf8fb-1411-0410-9b6e-c06b97aefaee

+++ tools/results.c
@@ -43,11 +43,11 @@ int main(int argc, char *argv[])
 	float response_time;
 	int tid;
 	time_t start_time = -1;
-	double total_response_time;
-	double ips;
+	float total_response_time;
+	float ips;
 	long long total_interaction_count = 0;
 	long long interaction_count[INTERACTION_TOTAL];
-	double interaction_response_time[INTERACTION_TOTAL];
+	float interaction_response_time[INTERACTION_TOTAL];
 
 	FILE *log_cpu;
 	FILE *log_io;
